handle logout
