+++ src/components/navbar/index.tsx
@@ -15,8 +15,24 @@ import { Button } from "@/components/ui/button";
 import { Link } from "react-router-dom";
 import assets from "@/assets/images";
 import { IoMenuSharp } from "react-icons/io5";
+import { useLogoutMutation } from "@/app/api/v0/auth";
+import { useAppSelector } from "@/app/hooks";
 
 export const Navbar = () => {
+  const [logout, { isLoading }] = useLogoutMutation();
+  const auth = useAppSelector((state) => state.authentication);
+  
+  const handleLogout = async () => {
+    try {
+      await logout().unwrap();
+      console.log("Logged out successfully");
+    } catch (err) {
+      console.error("Logout failed:", err);
+    } finally {
+      window.location.reload();
+    }
+  };
+
   return (
     <header className="shadow-sm">
       <div className="container flex h-16 w-full shrink-0 items-center justify-between px-4 md:px-6">
@@ -110,9 +126,15 @@ export const Navbar = () => {
           </SheetContent>
         </Sheet>
         <div className="flex items-center gap-4">
-          <Link to={"/auth/sign-in"}>
-            <Button>Sign In</Button>
-          </Link>
+          {auth.isAuthenticated ? (
+            <Button variant={"destructive"} onClick={handleLogout}>
+              {isLoading ? "Loading..." : "Logout"}
+            </Button>
+          ) : (
+            <Link to={"/auth/sign-in"}>
+              <Button>Sign In</Button>
+            </Link>
+          )}
         </div>
       </div>
     </header>
